* using size_t

+++ src/o_messages.h
@@ -72,11 +72,11 @@ public:
     int Activate(int No);
     int CanActivate(int Line);
     void NotifyPipe(int APipeId);
-    virtual void GetName(char *AName, int MaxLen);
-    virtual void GetInfo(char *AInfo, int MaxLen);
-    virtual void GetPath(char *APath, int MaxLen);
-    virtual void GetTitle(char *ATitle, int MaxLen, char *ASTitle, int SMaxLen);
-    virtual int GetRowLength(int ARow);
+    virtual void GetName(char *AName, size_t MaxLen);
+    virtual void GetInfo(char *AInfo, size_t MaxLen);
+    virtual void GetPath(char *APath, size_t MaxLen);
+    virtual void GetTitle(char *ATitle, size_t MaxLen, char *ASTitle, size_t SMaxLen);
+    virtual size_t GetRowLength(int ARow);
 
 
     int RunPipe(char *Dir, char *Command);
